# swagger.yaml
swagger: "2.0"
info:
  description: |
    An API used to create and query information about import jobs. This API is seperated into two parts;
    * The external interface is used by florence to create and query for information about an import job
    * The internal interface is used by services which are processing the import job. The services will use this API to update the state and add events to the job.
  version: "1.0.0"
  title: "ONS Import API"
  license:
    name: "Open Government Licence v3.0"
    url: "http://www.nationalarchives.gov.uk/doc/open-government-licence/version/3/"
basePath: "/v1"
tags:
- name: "External Interface"
- name: "Internal Interface"
schemes:
- "http"
parameters:
  instance_id:
    name: instance_id
    description: "A unique id for an instance"
    type: string
    in: path
    required: true
  job_id:
    name: job_id
    description: "A unique job id"
    type: string
    in: path
    required: true
  job:
    name: job
    in: body
    schema:
      $ref: '#/definitions/Job'
    required: true
  job_states:
    name: job_states
    description: "A list of comma separated job states to filter on. Eg created,submitted"
    type: array
    items:
      type: string
    in: query
  instance:
    name: instance
    description: "A state of an instance"
    schema:
      $ref: '#/definitions/Instance'
    in: body
    required: true
  file:
    name: File
    description: "A file to include in the job"
    in: body
    schema:
      $ref: '#/definitions/File'
    required: true
  event:
    name: event
    description: "An import event when publishing a dataset"
    in: body
    schema:
      $ref: '#/definitions/Event'
    required: true
  dimension_id:
    name: dimension_id
    description: "The name of a node (instanceId-Type.Value)"
    type: string
    in: path
    required: true
  value:
   name: value
   description: "A value to set within a type"
   type: string
   in: path
   required: true
  node_id:
   name: node_id
   description: "An unique node id"
   type: string
   in: path
   required: true
  inserted_observations:
    name: inserted_observations
    description: "A value to increment the inserted_observations within an instance"
    type: integer
    in: path
    required: true
securityDefinitions:
  FlorenceAPIKey:
    description: "API key used to allow florence users to create and query the progress of importing a dataset"
    type: apiKey
    in: header
    name: florence-token
  InternalAPIKey:
    description: "API key used to allow only internal services to update the state of an import job"
    type: apiKey
    in: header
    name: internal-token
paths:
  /jobs:
    get:
      tags:
      - "External Interface"
      summary: "Get a list of all jobs"
      description: "Lists can be filtered by the job state"
      produces:
       - "application/json"
      parameters:
      - $ref: '#/parameters/job_states'
      security:
      - FlorenceAPIKey: []
      responses:
        200:
          description: "A list of jobs has been returned"
          schema:
            $ref: '#/definitions/JobList'
        500:
          $ref: '#/responses/InternalError'
    post:
      tags:
      - "External Interface"
      summary: "Create an import job"
      description: |
        To import a dataset a job must be created first. To do this a data baker recipe is needed and the number of instances which the recipe creates. Once a job is created files can be added to the job and the state of the job can be changed.
      produces:
      - "application/json"
      parameters:
      - $ref: '#/parameters/job'
      security:
      - FlorenceAPIKey: []
      responses:
        201:
          description: "An import job was successfully created"
          schema:
            $ref: '#/definitions/Job'
        400:
          description: "Invalid json message was sent to the API"
        500:
          $ref: '#/responses/InternalError'
  /jobs/{job_id}:
    put:
      tags:
      - "External Interface"
      summary: "Update the jobs state"
      description: |
        Update the state of the job. If this is set to submitted, this shall trigger the
        import process.
      parameters:
      - $ref: '#/parameters/job_id'
      - $ref: '#/parameters/job'
      produces:
      - "application/json"
      security:
      - FlorenceAPIKey: []
      responses:
        200:
          description: "The job is in a queue"
        400:
          description: "Invalid json message was sent to the API"
        404:
          description: "JobId does not match any import jobs"
        500:
          $ref: '#/responses/InternalError'
    get:
        tags:
        - "External Interface"
        summary: "Get a job"
        description: "Get information about a single job"
        parameters:
         - $ref: '#/parameters/job_id'
        produces:
        - "application/json"
        security:
        - FlorenceAPIKey: []
        responses:
          200:
            description: "Return a single jobs information"
            schema:
              $ref: '#/definitions/Job'
          404:
            description: "JobId does not match any import jobs"
          500:
            $ref: '#/responses/InternalError'
  /jobs/{job_id}/files:
    put:
      tags:
      - "External Interface"
      summary: "Add a file into a job"
      description: "Add a file into a job, for each file added an alias name needs to be given. This name needs to link to the recipe"
      parameters:
      - $ref: '#/parameters/job_id'
      - $ref: '#/parameters/file'
      produces:
      - "application/json"
      security:
      - FlorenceAPIKey: []
      responses:
        200:
          description: "The file was added to the import job"
        400:
          description: "Invalid json message was sent to the API"
        404:
          description: "JobId does not match any import jobs"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}:
    get:
      tags:
      - "External Interface"
      summary: "Get Information about an instance"
      description: "Get the current state of an instance, this includes all events which have happened."
      parameters:
      - $ref: '#/parameters/instance_id'
      produces:
      - "application/json"
      security:
      - FlorenceAPIKey: []
      responses:
        200:
          description: "Return a single instance state"
          schema:
            $ref: '#/definitions/Instance'
        404:
          description: "InstanceId does not match any import jobs"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}/dimensions:
    get:
      tags:
      - "External Interface"
      summary: "Get all dimensions from an instance"
      description: "Get all dimensions from an instance"
      parameters:
      - $ref: '#/parameters/instance_id'
      produces:
      - "application/json"
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "Return a list of dimensions"
          schema:
            type: array
            items:
              $ref: '#/definitions/Dimension'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}/dimensions/{dimension_id}/options:
    get:
      tags:
      - "External Interface"
      summary: "Get all unique values from a dimension"
      description: "Get all unique values from a dimension"
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/dimension_id'
      produces:
      - "application/json"
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "Return a list of unique values"
          schema:
            type: object
            properties:
              dimension_id:
                description: "The name of the dimension"
                type: string
              values:
                description: "A list of all unique values within a dimension"
                type: array
                items:
                  type: string
        400:
          description: "instance_id does not match any instances"
        404:
          description: "dimension_id does not match any dimensions within the instance"
        500:
          $ref: '#/responses/InternalError'
    put:
      tags:
      - "Internal Interface"
      summary: "Update the number of observations within this instance"
      description: "Update the number of observations within this instance"
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/instance'
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "The event was added to the instance"
        400:
          description: "Invalid json message sent"
        401:
          $ref: '#/responses/UnauthorisedError'
        403:
          $ref: '#/responses/ForbiddenError'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}/events:
    post:
      tags:
       - "Internal Interface"
      summary: "Add an event to an instance"
      description: |
        Add a new event into an instance. Events can be for information or error messages.
        Each event must contain a type of event (Info or Error), a message to explain
        the event, time of the event and finally the message offset in kafka
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/event'
      security:
      - InternalAPIKey: []
      responses:
        201:
          description: "The event was added to the instance"
        400:
          description: "Invalid json message sent"
        401:
          $ref: '#/responses/UnauthorisedError'
        403:
          $ref: '#/responses/ForbiddenError'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}/inserted_observations/{inserted_observations}:
    put:
      tags:
      - "Internal Interface"
      summary: "Increment the observation inserted count"
      description: "This will add to the number already store in the api"
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/inserted_observations'
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "Added value to inserted observation"
        400:
          description: "Invalid value provided for inserted_observations"
        401:
          $ref: '#/responses/UnauthorisedError'
        403:
          $ref: '#/responses/ForbiddenError'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'


  /instances/{instance_id}/dimensions/{dimension_id}/options/{value}:
    put:
      tags:
      - "Internal Interface"
      summary: "Add a dimension for the instance"
      description: |
        Add a dimension found in the V4 data.
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/dimension_id'
      - $ref: '#/parameters/value'
      produces:
      - "application/json"
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "The dimension was added"
        401:
          $ref: '#/responses/UnauthorisedError'
        403:
          $ref: '#/responses/ForbiddenError'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'
  /instances/{instance_id}/dimensions/{dimension_id}/options/{value}/node_id/{node_id}:
    put:
      tags:
      - "Internal Interface"
      summary: "Update a dimension with the node_id"
      description: |
        Update the dimension with a nodeId
      parameters:
      - $ref: '#/parameters/instance_id'
      - $ref: '#/parameters/dimension_id'
      - $ref: '#/parameters/value'
      - $ref: '#/parameters/node_id'
      produces:
      - "application/json"
      security:
      - InternalAPIKey: []
      responses:
        200:
          description: "Updated the dimension with the nodeId"
        401:
          $ref: '#/responses/UnauthorisedError'
        403:
          $ref: '#/responses/ForbiddenError'
        404:
          description: "InstanceId does not match any instances"
        500:
          $ref: '#/responses/InternalError'
responses:
  InternalError:
    description: "Failed to process the request due to an internal error"
  ForbiddenError:
    description: "No header with a token key was provided"
  UnauthorisedError:
    description: "The token provided is unauthorised to carry out this operation"
definitions:
  JobList:
    description: "A list of import jobs"
    type: array
    items:
      $ref: '#/definitions/Job'
  Job:
    type: object
    description: "An object returned when an import job is created"
    properties:
      job_id:
        type: string
        readOnly: true
        description: "A unique id for an import job"
      recipe:
        type: string
        description: |
          The recipe to use when baking data. Below is a list of predefined recipes;
          * v4 - This will skip the data baking process.
      state:
        type: string
        description: |
          The state of the job, this can be;
           * created - The job has been created;
           * submitted - The job has been queue to be imported
           * completed - The job has been imported
           * error - The job was not imported (See the events in the instances)
      links:
        type: object
        readOnly: true
        properties:
          instance_ids:
            type: array
            description: "A list of URL to instances"
            items:
              type: string
      files:
        description: "A list of all files to be used in the job"
        type: array
        items:
          $ref: '#/definitions/File'
  Instance:
    type: object
    properties:
      instance_id:
        type: string
        readOnly: true
        description: "A unique id for a job"
      job_url:
        description: "A URL to the job which this instance is contained in"
        type: string
        example: "http://localhost:21800/jobs/042e216a-7822-4fa0-a3d6-e3f5248ffc35"
      state:
        type: string
        description: |
          The state of the import job
      events:
        type: array
        readOnly: true
        description: "A list of events took place for this job"
        items:
          $ref: '#/definitions/Event'
      number_of_observations:
        type: integer
        description: "The number of observations in this instance"
      headers:
        type: array
        description: "The header information from a V4 file"
        items:
          type: string
      last_updated:
        type: string
        readOnly: true
        description: "The last time an event happened"
  Event:
    type: object
    properties:
      type:
        type: string
        description: |
          The type of event, this can be;
          * Info - for an information event
          * Error - for an error event
      service:
        type: string
        description: "The name of the service which triggered the event"
      time:
        type: string
        description: "The time of the event in ISO 8601 format"
      message:
        type: string
        description: "A description of the event"
      message_offset:
        type: string
        description: "The kafka message offset (Used to locate the kafka message)"
  Dimension:
    type: object
    description: "A dimension object"
    properties:
      dimension_id:
        description: "The name of the dimension (instanceId-dimension.value)."
        type: string
      value:
        description: "The value of the dimension"
        type: string
      node_id:
        description: "The id of the node"
        type: string
  File:
    type: object
    properties:
      alias_name:
        description: "An alias name for the file"
        type: string
      url:
        description: "The full S3 path including zone, and bucket"
        type: string